absensi/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
mysql-connector-python==8.1.0
tkinter
customtkinter==5.2.0
python-dotenv==1.0.0
requests==2.31.0
dlib==19.24.2
